companies

+++ FrontEnd/AH.Web/Server/Services/DomainService.cs
@@ -1,4 +1,3 @@
-using System.Net.Http.Headers;
 using AH.Metadata.Shared.V1.Models.Responses.Domains;
 using AH.Web.Server.Services.Interfaces;
 
@@ -7,22 +6,21 @@ namespace AH.Web.Server.Services;
 public class DomainService : IDomainService
 {
     private readonly HttpClient _httpClient;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
     public DomainService(HttpClient httpClient, IHttpContextAccessor httpContextAccessor)
     {
         _httpClient = httpClient;
-        var token = httpContextAccessor.HttpContext?.Request.Headers["Authorization"].ToString() ?? string.Empty;
-        _httpClient.DefaultRequestHeaders.Authorization =
-            new AuthenticationHeaderValue("Bearer", token.Replace("Bearer ", string.Empty));
+        _httpContextAccessor = httpContextAccessor;
     }
-    public async Task<List<DomainResponse>> GetDomains() => await _httpClient.Get<DomainResponse>("v1/domains");
+    public async Task<List<DomainWithCompaniesResponse>> GetDomains() => await _httpClient.Get<DomainWithCompaniesResponse>("v1/domains", _httpContextAccessor);
 
-    public async Task<DomainResponse> UpdateDomain(DomainResponse domain)
-        => await _httpClient.Upsert<DomainResponse>($"v1/domains/{domain.UId}", domain, true);
+    public async Task<DomainWithCompaniesResponse> UpdateDomain(DomainWithCompaniesResponse domain)
+        => await _httpClient.Upsert<DomainWithCompaniesResponse>($"v1/domains/{domain.UId}", domain, true, _httpContextAccessor);
 
-    public async Task<DomainResponse> CreateDomain(DomainResponse domain) =>
-         await _httpClient.Upsert<DomainResponse>($"v1/domains/{domain.UId}", domain, false);
+    public async Task<DomainWithCompaniesResponse> CreateDomain(DomainWithCompaniesResponse domain) =>
+         await _httpClient.Upsert<DomainWithCompaniesResponse>($"v1/domains", domain, false, _httpContextAccessor);
 
-    public async Task DeleteDomain(Guid uid) => await _httpClient.Delete($"v1/domains/{uid}");
+    public async Task DeleteDomain(Guid uid) => await _httpClient.Delete($"v1/domains/{uid}", _httpContextAccessor);
     
 }
