+++ FrontEnd/AH.Web/Server/Services/MasterTagCategoryService.cs
@@ -1,5 +1,4 @@
-﻿using System.Net.Http.Headers;
-using AH.Metadata.Shared.V1.Models.Responses.MasterTagCategories;
+﻿using AH.Metadata.Shared.V1.Models.Responses.MasterTagCategories;
 using AH.Web.Server.Services.Interfaces;
 
 namespace AH.Web.Server.Services;
@@ -7,32 +6,31 @@ namespace AH.Web.Server.Services;
 public class MasterTagCategoryService : IMasterTagCategoryService
 {
     private readonly HttpClient _httpClient;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
     public MasterTagCategoryService(HttpClient httpClient, IHttpContextAccessor httpContextAccessor)
     {
         _httpClient = httpClient;
-        var token = httpContextAccessor.HttpContext?.Request.Headers["Authorization"].ToString() ?? string.Empty;
-        _httpClient.DefaultRequestHeaders.Authorization =
-            new AuthenticationHeaderValue("Bearer", token.Replace("Bearer ", string.Empty));
+        _httpContextAccessor = httpContextAccessor;
     }
 
     public async Task<List<MasterTagCategoryResponse>> GetMasterTagCategories() =>
-        await _httpClient.Get<MasterTagCategoryResponse>("v1/mastertagcategories");
+        await _httpClient.Get<MasterTagCategoryResponse>("v1/mastertagcategories", _httpContextAccessor);
 
     public async Task<MasterTagCategoryResponse> CreateMasterTagCategory(MasterTagCategoryResponse category)
     {
-        var response =  await _httpClient.Upsert<MasterTagCategoryResponse>
-                ($"v1/mastertagcategories?Name={category.Name}&ClassName={category.ClassName}&Type={category.Type}",
-                    category, false);
+        var response = await _httpClient.Upsert<MasterTagCategoryResponse>
+        ($"v1/mastertagcategories?Name={category.Name}&ClassName={category.ClassName}&Type={category.Type}",
+            category, false, _httpContextAccessor);
         return response;
     }
-        
+
 
     public async Task<MasterTagCategoryResponse> UpdateMasterTagCategory(MasterTagCategoryResponse category) =>
         await _httpClient.Upsert<MasterTagCategoryResponse>
         ($"v1/mastertagcategories/{category.UId}?Name={category.Name}&ClassName={category.ClassName}&Type={category.Type}",
-            category, true);
+            category, true, _httpContextAccessor);
 
     public async Task DeleteMasterTagCategory(Guid uid) =>
-        await _httpClient.Delete($"v1/mastertagcategories/{uid}");
+        await _httpClient.Delete($"v1/mastertagcategories/{uid}", _httpContextAccessor);
 }
